+++ pkg/eyecandy/emoji.go
@@ -1,0 +1,26 @@
+package eyecandy
+
+import (
+	"fmt"
+	"github.com/kyokomi/emoji/v2"
+	"regexp"
+)
+
+func ESPrintf(emojisDisabled bool, format string, v ...interface{}) string {
+	if emojisDisabled {
+		return fmt.Sprintf(removeEmojiFromString(format), v...)
+	}
+	return emoji.Sprintf(format, v...)
+}
+
+func ESPrint(emojisDisabled bool, s string) string {
+	if emojisDisabled {
+		return fmt.Sprint(removeEmojiFromString(s))
+	}
+	return emoji.Sprint(s)
+}
+
+func removeEmojiFromString(s string) string {
+	re := regexp.MustCompile(`:[a-zA-Z0-9-_+]+?:`)
+	return re.ReplaceAllString(s, "")
+}
